fix: Fable build stopped working when "adaptify" was added
1. Remove Adaptify.MSBuild (because that doesn't work with the Fable build, and we don't want separate ways of working for the .NET and Fable builds)
2. Temporarily downgrade to .NET 6 so that the "adaptify" dotnet tool works
3. Add "Directory.Build.targets" containing a custom target to run Adaptify in "local" mode and add the generated files to the project file
4. Change the "test+watch" script in "package.json" to explicitly run the custom target when necessary in addition to running both Fable and Mocha in watch mode.

For the .NET build, the custom target is run automatically because it specifies that it should run before "CoreCompile".

However, for the Fable build the 4th step is necessary because it does *NOT* run the the custom target automatically - Fable has it's own custom parsing of project files etc. to work out the files to compile rather than using msbuild.