fix: specify correct number of items to retain
It's necessary to add 1 to the position when retaining because it's the number of items to retain which needs to be specified, rather than the highest index.

For example, if you want to retain items with indices 0-3 you need to specify "retain 4 items".

This commit also corrects `toIndexListDelta` in "tests/Ylmish.Tests/Y.Delta.fs" by ensuring the first Index is set to Index.Zero.

With these changes, all the `Y.Delta` tests pass.